split the scroll logic
