+++ tests/Table.test.jsx
@@ -62,7 +62,7 @@ const rowData = [
 
 const NeetoUITable = props => (
   <BrowserRouter>
-    <Table {...props} scroll={{ x: 1000 }} />
+    <Table {...props} />
   </BrowserRouter>
 );
 
@@ -80,7 +80,7 @@ describe("Table", () => {
   });
 
   it("should render all the rows", () => {
-    render(<NeetoUITable {...{ columnData, rowData }} virtual={false} />);
+    render(<NeetoUITable {...{ columnData, rowData }} />);
     const row = screen.getAllByRole("row");
     expect(row.length).toBe(4);
   });
@@ -150,7 +150,6 @@ describe("Table", () => {
         {...{ columnData, rowData }}
         defaultPageSize={2}
         shouldDynamicallyRenderRowSize={false}
-        virtual={false}
       />
     );
     const row = screen.getAllByRole("row");
